requirements.txt
pandas
fastapi
pyarrow
python-dotenv
uvicorn
sqlachemy
psycopg2-binary
